bring fetching reward balances back to ss Store, but no Promise all
