+++ Frontend-v1-Original/stores/connectors/viem.ts
@@ -1,5 +1,12 @@
-import { createPublicClient, http, fallback } from "viem";
+import {
+  createPublicClient,
+  http,
+  fallback,
+  ContractFunctionConfig,
+  MulticallParameters,
+} from "viem";
 import { canto } from "viem/chains";
+import { CONTRACTS } from "../constants/constants";
 
 const dexvaults = http("https://canto.dexvaults.com");
 const plexnode = http("https://mainnode.plexnode.org:8545");
@@ -20,5 +27,40 @@ const client = createPublicClient({
     }
   ),
 });
+
+export function chunkArray<T>(array: T[], chunkSize: number): T[][] {
+  const result = [];
+  for (let i = 0; i < array.length; i += chunkSize) {
+    const chunk = array.slice(i, i + chunkSize);
+    result.push(chunk);
+  }
+  return result;
+}
+
+export async function multicallChunks<
+  TContracts extends ContractFunctionConfig[],
+  TAllowFailure extends boolean = true
+>(chunks: MulticallParameters<TContracts, TAllowFailure>["contracts"][]) {
+  let result = [];
+
+  if (chunks.length === 1) {
+    const chunkResult = await client.multicall({
+      allowFailure: false,
+      multicallAddress: CONTRACTS.MULTICALL_ADDRESS,
+      contracts: chunks[0],
+    });
+    return chunkResult;
+  }
+
+  for (const chunk of chunks) {
+    const chunkResult = await client.multicall({
+      allowFailure: false,
+      multicallAddress: CONTRACTS.MULTICALL_ADDRESS,
+      contracts: chunk,
+    });
+    result.push(chunkResult);
+  }
+  return result.flat();
+}
 
 export default client;
